requirements.txt
alembic==1.7.5
aniso8601==9.0.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.2
Flask-Admin==1.5.8
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-MySQL==1.5.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
Markdown==3.4.1
MarkupSafe==2.0.1
mysql-connector-python==8.2.0
phonenumbers==8.13.26
Pillow==9.0.0
protobuf==4.21.12
pycparser==2.21
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.29
urllib3==1.26.9
Werkzeug==2.0.2
WTForms==3.0.1
zipp==3.8.0
